+++ resources/views/components/register-form.blade.php
@@ -1,4 +1,6 @@
 <div class="authkit-form-container">
+    @include('authkit::components.authkit-styles')
+
     <form method="POST" action="{{ $action }}" class="authkit-form">
         @csrf